feat: add `generate-genesis` subcommand to tempo binary
Extracts genesis generation logic from xtask into a shared `tempo-genesis`
crate and adds `tempo generate-genesis` as a first-class subcommand.

Amp-Thread-ID: https://ampcode.com/threads/T-019dab30-dc50-7555-9e76-08dd7df121ae
