+++ crates/genesis/Cargo.toml
@@ -1,5 +1,6 @@
 [package]
-name = "tempo-xtask"
+name = "tempo-genesis"
+description = "Genesis generation for Tempo"
 
 version.workspace = true
 edition.workspace = true
@@ -13,44 +14,31 @@ workspace = true
 [dependencies]
 tempo-chainspec.workspace = true
 tempo-contracts.workspace = true
-tempo-primitives.workspace = true
 tempo-commonware-node-config.workspace = true
-tempo-genesis.workspace = true
-tempo-validator-config.workspace = true
 tempo-dkg-onchain-artifacts.workspace = true
 tempo-evm.workspace = true
 tempo-precompiles.workspace = true
-tempo-revm.workspace = true
+tempo-validator-config.workspace = true
 
 alloy = { workspace = true, features = [
   "genesis",
-  "providers",
-  "reqwest",
-  "reqwest-rustls-tls",
   "signers",
   "signer-local",
   "signer-mnemonic",
   "rand",
-  "sol-types",
   "getrandom",
 ] }
-alloy-primitives = { workspace = true, features = ["asm-keccak"] }
-coins-bip32.workspace = true
+alloy-primitives.workspace = true
 clap = { workspace = true, features = ["derive"] }
 commonware-codec.workspace = true
 commonware-consensus.workspace = true
 commonware-cryptography.workspace = true
 commonware-math.workspace = true
 commonware-utils.workspace = true
-const-hex.workspace = true
 eyre.workspace = true
 indicatif = { workspace = true, features = ["rayon"] }
 itertools.workspace = true
 rand_08.workspace = true
 rayon.workspace = true
 reth-evm.workspace = true
-reth-network-peers = { workspace = true, features = ["secp256k1"] }
-secp256k1 = { workspace = true, features = ["rand"] }
-serde.workspace = true
 serde_json.workspace = true
-tokio.workspace = true
