+++ xtask/src/generate_genesis.rs
@@ -25,67 +25,6 @@ impl GenerateGenesis {
             .await
             .wrap_err("failed generating genesis")?;
 
-        let json =
-            serde_json::to_string_pretty(&genesis).wrap_err("failed encoding genesis as JSON")?;
-
-        std::fs::create_dir_all(&output).wrap_err_with(|| {
-            format!(
-                "failed to create directory and parents for `{}`",
-                output.display()
-            )
-        })?;
-        let genesis_dst = output.join("genesis.json");
-        std::fs::write(&genesis_dst, json).wrap_err_with(|| {
-            format!("failed writing genesis to file `{}`", genesis_dst.display())
-        })?;
-
-        if let Some(consensus_config) = consensus_config {
-            println!(
-                "consensus config generated for `{}` validators; writing to disk...",
-                consensus_config.validators.len()
-            );
-            for validator in consensus_config.validators {
-                std::fs::create_dir_all(validator.dst_dir(&output)).wrap_err_with(|| {
-                    format!(
-                        "failed creating target directory to store validator specifici keys at `{}`",
-                        validator.dst_dir(&output).display()
-                    )
-                })?;
-                let signing_key_dst = validator.dst_signing_key(&output);
-                std::fs::File::create(&signing_key_dst)
-                    .map_err(eyre::Report::new)
-                    .and_then(|f| {
-                        validator
-                            .signing_key
-                            .to_writer(f)
-                            .map_err(eyre::Report::new)
-                    })
-                    .wrap_err_with(|| {
-                        format!(
-                            "failed writing ed25519 signing key to `{}`",
-                            signing_key_dst.display()
-                        )
-                    })?;
-                let signing_share_dst = validator.dst_signing_share(&output);
-                validator
-                    .signing_share
-                    .write_to_file(&signing_share_dst)
-                    .wrap_err_with(|| {
-                        format!(
-                            "failed writing bls12381 signing share to `{}`",
-                            signing_share_dst.display()
-                        )
-                    })?;
-                println!(
-                    "validator keys written to `{}`, `{}`",
-                    signing_key_dst.display(),
-                    signing_share_dst.display()
-                );
-            }
-        } else {
-            println!("no consensus config generated; likely didn't provide --validators flag");
-        }
-
-        Ok(())
+        tempo_genesis::write_genesis(&output, &genesis, consensus_config.as_ref())
     }
 }
